edit works

+++ my-app/app/job/page.js
@@ -1,3 +1,4 @@
+
 'use client'
 import React, { useState, useEffect } from 'react';
 import Head from 'next/head';
@@ -7,9 +8,13 @@ export default function Jobs() {
   const [company, setCompany] = useState('');
   const [position, setPosition] = useState('');
   const [job,setJob] = useState([])
-  const [jobs, setJobs] = useState([]); // This state will keep track of the jobs
+  const [jobs, setJobs] = useState([]); 
   const [error, setError] = useState('');
   const [loading, setLoading] = useState(true);
+  const [deleteMessage, setDeleteMessage] = useState('');
+  const [isEditModalVisible, setIsEditModalVisible] = useState(false);
+  const [currentEditingJob, setCurrentEditingJob] = useState(null);
+  
 
   useEffect(() => {
     // Fetch jobs when the component mounts
@@ -29,7 +34,7 @@ export default function Jobs() {
             }
           });
           setJobs(response.data.jobs);
-        //   console.log(jobs)
+       
         //   console.log(Array.isArray(jobs))
         } catch (err) {
           setError('Failed to fetch jobs. Please make sure you are logged in.');
@@ -58,7 +63,7 @@ export default function Jobs() {
       }
     );
       
-          console.log("r",response.data.job)
+       console.log("r")
           const newJob = response.data.job;
          setJobs(currentJobs => [...currentJobs, newJob]); // Append new job to the current list
 
@@ -73,34 +78,80 @@ export default function Jobs() {
         }
       };
 
-      const handleDelete = async (jobId) => {
+    const handleDelete = async (jobId) => {
+        // Display confirmation dialog
+        const isConfirmed = window.confirm('Are you sure you want to delete this job?');
+      
+        // Proceed only if the user confirmed
+        if (isConfirmed) {
+          try {
+            const token = localStorage.getItem('token');
+            const url = `http://localhost:8080/api/jobs/${jobId}`;
+      
+            await axios.delete(url, {
+              headers: {
+                Authorization: `Bearer ${token}`
+              }
+            });
+      
+            // Update state to reflect the deleted job
+            setJobs(currentJobs => currentJobs.filter(job => job._id !== jobId));
+      
+            // Set a success message
+            setDeleteMessage('Job deleted successfully.');
+      
+            // Optionally, clear the message after a few seconds
+            setTimeout(() => setDeleteMessage(''), 500);
+          } catch (error) {
+            console.error('Error deleting job:', error);
+            if (error.response) {
+              console.error('Server responded with status code:', error.response.status);
+            }
+            // Set an error message if needed
+            setDeleteMessage('Failed to delete the job.');
+          }
+        } else {
+          // User canceled the action, you can optionally handle this case
+          console.log('Job deletion canceled by the user.');
+        }
+      };
+
+      const handleEditClick = (job) => {
+        setCurrentEditingJob(job);
+        setIsEditModalVisible(true);
+      };
+
+      const handleEditSubmit = async (e) => {
+        e.preventDefault();
+        if (!currentEditingJob) return;
+      
         try {
           const token = localStorage.getItem('token');
-          const url = `http://localhost:8080/api/jobs/${jobId}`;
-          console.log('Attempting to delete job with URL:', url); // Debug: Check the constructed URL
-      
-          await axios.delete(url, {
+          const response = await axios.patch(`http://localhost:8080/api/jobs/${currentEditingJob._id}`, currentEditingJob, {
             headers: {
               Authorization: `Bearer ${token}`
             }
           });
       
-          // If this line is reached, deletion was successful
-          setJobs(currentJobs => currentJobs.filter(job => job.id !== jobId));
+          // Update the jobs list with the updated job
+          setJobs(jobs.map(job => job._id === currentEditingJob._id ? response.data.job : job));
+          setIsEditModalVisible(false); // Close the modal
+          setCurrentEditingJob(null); // Reset current editing job
         } catch (error) {
-          console.error('Error deleting job:', error);
-          if (error.response) {
-            console.error('Server responded with status code:', error.response.status);
-          }
+          console.error('Error updating job:', error);
+          // Optionally, handle errors (e.g., display error message)
         }
       };
       
       
       
+      
+      
+      
 
 
-  console.log('Current jobs:', jobs);
-  console.log('Number of jobs:', jobs.length);
+//   console.log('Current jobs:', jobs);
+//   console.log('Number of jobs:', jobs.length);
   return (
     <div className="min-h-screen bg-gray-100">
       <Head>
@@ -112,6 +163,60 @@ export default function Jobs() {
       </header>
       <main className="container mx-auto mt-10">
         <div className="bg-white p-8 rounded-lg shadow-md">
+        {
+            isEditModalVisible && (
+                <div className="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full" id="my-modal">
+                <div className="relative top-20 mx-auto p-5 border w-96 shadow-lg rounded-md bg-white">
+                    <h2 className="text-2xl font-bold text-center mb-8">Edit Job</h2>
+                    <form onSubmit={handleEditSubmit}>
+                    <div className="mb-4">
+                        <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="edit-company">
+                        Company
+                        </label>
+                        <input
+                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                        id="edit-company"
+                        type="text"
+                        placeholder="Company"
+                        value={currentEditingJob?.company || ''}
+                        onChange={(e) => setCurrentEditingJob({ ...currentEditingJob, company: e.target.value })}
+                        />
+                    </div>
+                    <div className="mb-6">
+                        <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="edit-position">
+                        Position
+                        </label>
+                        <input
+                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                        id="edit-position"
+                        type="text"
+                        placeholder="Position"
+                        value={currentEditingJob?.position || ''}
+                        onChange={(e) => setCurrentEditingJob({ ...currentEditingJob, position: e.target.value })}
+                        />
+                    </div>
+                    <div className="flex justify-center">
+                        <button
+                        className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
+                        type="submit"
+                        >
+                        Save Changes
+                        </button>
+                        <button
+                        onClick={() => setIsEditModalVisible(false)}
+                        className="ml-4 bg-gray-500 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
+                        type="button"
+                        >
+                        Cancel
+                        </button>
+                    </div>
+                    </form>
+                </div>
+                </div>
+            )
+        }
+
+
           <h2 className="text-2xl font-bold text-center mb-8">New Job</h2>
           <form onSubmit={handleJobSubmit}>
             <div className="mb-4">
@@ -151,34 +256,41 @@ export default function Jobs() {
           </form>
         </div>
         <div className="mt-10 bg-white p-8 rounded-lg shadow-md">
+        {deleteMessage && (
+      <div className="text-center my-4 text-red-500">
+        {deleteMessage}
+      </div>
+    )}
         {jobs.length !== 0 ? (
-  jobs.map((job, index) => (
-    <div key={index} className="bg-white p-4 rounded-lg shadow-md mb-4 flex justify-between items-center">
-      <div>
-      <div className="p-2 rounded"> 
-      <p className="text-lg font-bold">{job.position}</p>
-      </div>
-      <div className="bg-gray-200 p-2 rounded"> 
-        <p className="text-lg font-bold">{job.company}</p>
-      </div>
-     
-      </div>
-      <div className="flex justify-center items-center">
-        <p className="text-xs bg-purple-200 text-purple-700 px-2 py-1 rounded-full">{new Date(job.updatedAt).toLocaleString('en-US', { 
-          dateStyle: 'medium', 
-          timeStyle: 'short' 
-        })}</p>
-        <p className="text-xs bg-yellow-200 text-yellow-700 px-2 py-1 rounded-full ml-2">{job.status}</p>
-      </div>
-      <div className="flex">
-        <button className="text-blue-600 hover:text-blue-800 mr-2">Edit</button>
-        <button className="text-red-600 hover:text-red-800" onClick={() => handleDelete(job._id)}>Delete</button>
-      </div>
-    </div>
-  ))
-) : (
-  <p className="text-center text-gray-700">You have no jobs to display</p>
-)}
+            jobs.map((job, index) => (
+                <div key={index} className="bg-white p-4 rounded-lg shadow-md mb-4 flex justify-between items-center">
+                <div>
+                <div className="p-2 rounded"> 
+                <p className="text-lg font-bold">{job.position}</p>
+                </div>
+                <div className="bg-gray-200 p-2 rounded"> 
+                    <p className="text-lg font-bold">{job.company}</p>
+                </div>
+                
+                </div>
+                <div className="flex justify-center items-center">
+                    <p className="text-xs bg-purple-200 text-purple-700 px-2 py-1 rounded-full">{new Date(job.updatedAt).toLocaleString('en-US', { 
+                    dateStyle: 'medium', 
+                    timeStyle: 'short' 
+                    })}</p>
+                    <p className="text-xs bg-yellow-200 text-yellow-700 px-2 py-1 rounded-full ml-2">{job.status}</p>
+                </div>
+                <div className="flex">
+                    <button className="text-blue-600 hover:text-blue-800 mr-2" onClick={() => handleEditClick(job)}>Edit</button>
+                    <button className="text-red-600 hover:text-red-800" onClick={() => handleDelete(job._id)}>Delete</button>
+                    
+                </div>
+                </div>
+            ))
+            ) : (
+            <p className="text-center text-gray-700">You have no jobs to display</p>
+            )
+        }
 
         </div>
       </main>
